+++ ember-addon-test/vite.config.mts
@@ -1,0 +1,16 @@
+import { defineConfig } from 'vite';
+import { plugins } from './vite-plugin.mjs';
+
+
+// https://vitejs.dev/config/
+export default defineConfig({
+  optimizeDeps: {
+    include: [
+      // path.join(__dirname, '**/*'),
+      // process.cwd(),
+    ]
+  },
+  plugins: [
+    ...plugins(),
+  ],
+})
