+++ ember-resources/index.html
@@ -5,7 +5,7 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vite + TS</title>
 
-    <link rel="stylesheet" href="node_modules/ember-addon-test/node_modules/qunit/qunit/qunit.css" />
+    <link rel="stylesheet" href="/qunit.css" />
   </head>
   <body>
